move parameter space to utils
