+++ pgdrive/scene_creator/blocks/t_intersection.py
@@ -1,10 +1,9 @@
 from pgdrive.constants import Goal
 from pgdrive.pg_config.parameter_space import Parameter, BlockParameterSpace
-from pgdrive.pg_config.pg_space import PGSpace
+from pgdrive.utils.pg_space import PGSpace
 from pgdrive.scene_creator.blocks.intersection import InterSection
 from pgdrive.scene_creator.lane.abs_lane import LineType, LineColor
 from pgdrive.scene_creator.road.road import Road
-from pgdrive.constants import Decoration
 from pgdrive.scene_creator.blocks.block import BlockSocket
