+++ pgdrive/utils/element.py
@@ -5,7 +5,7 @@ from typing import Dict
 from panda3d.bullet import BulletWorld
 from panda3d.core import NodePath
 from pgdrive.pg_config import PGConfig
-from pgdrive.pg_config.pg_space import PGSpace
+from pgdrive.utils.pg_space import PGSpace
 from pgdrive.utils.asset_loader import AssetLoader
 from pgdrive.world.pg_physics_world import PGPhysicsWorld
